Add DatabaseManagerAuthSettings extension to hold any method that interacts with AuthSettings without having to AuthSettings and its dependency chain to th Share Extension target
